app/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.26.0
google-generativeai>=0.5.0